refactor(orders): dissolve order-admin.service into concern modules
Split the residual 802-line order-admin.service.ts into single-concern
modules and delete it:

- order-queue.service.ts   OrderService reads + work-axis (start, handler,
                           status) + queue helpers
- order-photo.service.ts   the 5 photo handlers
- order-payment.service.ts updateOrderPayment
- order.service.ts         gains getOrderDetailById (Order-aggregate read)
- order.repository.ts      gains getOrderServiceOrThrow (+ prepared stmt),
                           the guard shared across queue and photo

Pure relocation, no behavior change: same validations, transactions,
permission gates. routes/admin/orders.ts repointed (route bodies unchanged);
order-admin.schema.ts kept (package-exported). Completes the #1/#5/#6/#2
teardown of the original 1506-line god-file.

42/42 tests pass, type-check + biome clean.
